+++ buession-springboot-web/src/main/java/com/buession/springboot/web/reactive/autoconfigure/ReactiveWebFluxConfiguration.java
@@ -29,16 +29,35 @@ import com.buession.web.reactive.filter.RequestContextFilter;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.AutoConfigureOrder;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
+import org.springframework.boot.autoconfigure.http.codec.CodecsAutoConfiguration;
+import org.springframework.boot.autoconfigure.validation.ValidationAutoConfiguration;
+import org.springframework.boot.autoconfigure.web.reactive.ReactiveMultipartAutoConfiguration;
+import org.springframework.boot.autoconfigure.web.reactive.ReactiveWebServerFactoryAutoConfiguration;
+import org.springframework.boot.autoconfigure.web.reactive.WebFluxAutoConfiguration;
+import org.springframework.boot.autoconfigure.web.reactive.WebSessionIdResolverAutoConfiguration;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
+import org.springframework.core.Ordered;
 import org.springframework.core.ReactiveAdapterRegistry;
+import org.springframework.web.reactive.config.WebFluxConfigurationSupport;
+import org.springframework.web.reactive.config.WebFluxConfigurer;
 
 /**
  * @author Yong.Teng
  */
-@AutoConfiguration
+@AutoConfiguration(after = {ReactiveWebServerFactoryAutoConfiguration.class, CodecsAutoConfiguration.class,
+		ReactiveMultipartAutoConfiguration.class, ValidationAutoConfiguration.class,
+		WebSessionIdResolverAutoConfiguration.class})
 @ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.REACTIVE)
+@ConditionalOnClass(WebFluxConfigurer.class)
+@ConditionalOnMissingBean({WebFluxConfigurationSupport.class})
+@ConditionalOnBean({WebFluxAutoConfiguration.WebFluxConfig.class})
+@AutoConfigureOrder(Ordered.HIGHEST_PRECEDENCE + 10)
 public class ReactiveWebFluxConfiguration extends WebFluxConfiguration {
 
 	public ReactiveWebFluxConfiguration(ConfigurableApplicationContext applicationContext,